Accept \left. and \right. as null delimiters.
When finding \parindent (do-indent, do-plain-item, do-textindent), use
the-dimen, not find-dimen!

When \TZPsinglepage: (1) Embed basic CSS style within document instead of
generating CSS file. (2) Embed images as data URIs instead of pointing to
files. This allows for single-page HTMLs to be self-sufficient when emailing to
a Kindle.

do-end-equation: Make sure correct environment name is identified and matched.

$$: Allow for single $ to occur even when not nested instead braces. This is
because {tabular} allows this situation.